client: isolated into a separated module
