+++ mypy.ini
@@ -143,6 +143,9 @@ ignore_errors = True
 [mypy-karapace.utils]
 ignore_errors = True
 
+[mypy-karapace.client]
+ignore_errors = True
+
 [mypy-karapace.rapu]
 ignore_errors = True
