+++ tests/integration/test_schema_protobuf.py
@@ -4,8 +4,8 @@ karapace - schema tests
 Copyright (c) 2019 Aiven Ltd
 See LICENSE for details
 """
+from karapace.client import Client
 from karapace.protobuf.kotlin_wrapper import trim_margin
-from karapace.utils import Client
 from tests.utils import create_subject_name_factory
 
 import logging
